Add files via upload
Created an object pool design pattern which creates an object of type DatabaseObject. There are two concrete pool object implementation, Fixed and dynamic size pool. Both are thread safe singleton pattern. There is a factory class which will create the instance of DatabaseObject and return it to pool class. The fixed size pool implementation is incomplete. Once the max size is reached, then the user has to wait for objects. This part is not implemented as of now.