Add files via upload
Observer implementation in c++. Here server is a concrete observable and Clients are concrete observers. While creating a client, i am passing the reference of server to which this client is subscribed. This way whenever there is a change in server and clients gets notified...client can read the value from server to which it is subscribed.(because of reference which is passed earlier).